Arreglado error out of index. Quité los time.sleep(), ahora ejecuta muchisimo más rápido. Cambié el timer y la BD para que mida en milisegundos

+++ static/CreateSQL.sql
@@ -32,7 +32,7 @@ VALUES
 CREATE TABLE IF NOT EXISTS Games (
 idGames INT NOT NULL AUTO_INCREMENT,
 victoria TINYINT NOT NULL,
-duracion INT NOT NULL,
+duracion FLOAT NOT NULL,
 movimientos INT NOT NULL,
 Mazo VARCHAR(5000) NOT NULL,
 Estrategia_idEstrategia INT NOT NULL,
